requirements.txt
python-telegram-bot==20.7
openai>=1.0.0
python-dotenv==1.0.0
APScheduler==3.10.4
psycopg2-binary==2.9.9
